% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVs.R
\name{CVs}
\alias{CVs}
\title{A Cultural Value Index (CV) Function}
\source{
Reyes-García, V., T. Huanca, V. Vadez, W. Leonard, and D. Wilkie. 2006. Cultural, Practical, and Economic Value of Wild Plants: A Quantitative Study in the Bolivian Amazon. Economic Botany 60(1):62–74.
}
\usage{
CVs(data)
}
\arguments{
\item{data}{is an ethnobotany data set with column 1 'informant' and 2 'sp_name' as row identifiers of informants and of species names respectively.
The rest of the columns are the identified ethnobotany use categories. The data should be populated with counts of uses per person (should be 0 or 1 values).}
}
\description{
This function allows you to calculate the Cultural Value Index (CV) published by Reyes-García et al. (2006).
}
\examples{
CVs(ethnobotanydata)

}
\keyword{cultural}
\keyword{ethnobotany,}
\keyword{value}
