% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethno_alluvial.R
\name{ethno_alluvial}
\alias{ethno_alluvial}
\title{Alluvial plot of ethnobotany uses and species}
\usage{
ethno_alluvial(data)
}
\arguments{
\item{data}{is an ethnobotany data set with column 1 'informant' and 2 'sp_name' as row identifiers of informants and of species names respectively.
The rest of the columns are the identified ethnobotany use categories. The data should be populated with counts of uses per person (should be 0 or 1 values).}
}
\value{
Alluvial diagram figure for each use (top half) related to each 'sp_name' (bottom half) in the data set. 
To change variable names try using the dplyr rename function.
}
\description{
Creates a simple alluvial plot of species and uses for ethnobotany studies.
}
\section{Warning}{


Identification for informants and species must be listed by the names 'informant' and 'sp_name' respectively in the data set.
The rest of the columns represent the identified ethnobotany use categories. These data should be populated with counts of uses per person (should be 0 or 1 values).
}

\examples{

#Use built-in ethnobotany data example
ethno_alluvial(ethnobotanydata)

#Generate random dataset of three informants uses for four species

eb_data <- data.frame(replicate(10,sample(0:1,20,rep=TRUE)))
names(eb_data) <- gsub(x = names(eb_data), pattern = "X", replacement = "Use_")  
eb_data$informant <- sample(c('User_1', 'User_2', 'User_3'), 20, replace=TRUE)
eb_data$sp_name <- sample(c('sp_1', 'sp_2', 'sp_3', 'sp_4'), 20, replace=TRUE)

ethno_alluvial(eb_data)

}
\keyword{arith}
\keyword{graphs}
\keyword{logic}
\keyword{math}
\keyword{methods}
\keyword{misc}
\keyword{survey}
