% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnoChord.R
\name{ethnoChord}
\alias{ethnoChord}
\title{Create a chord diagram of ethnobotany uses and species}
\usage{
ethnoChord(data)
}
\arguments{
\item{data}{is an ethnobotany data set with column 1 'informant' and 2 'sp_name' as row identifiers of informants and of species names respectively.
The rest of the columns are the identified ethnobotany use categories. The data should be populated with counts of uses per person (should be 0 or 1 values).}
}
\description{
This function allows you to create a chord diagram of species and uses for ethnobotany studies.
}
\examples{

ethnoChord(ethnobotanydata)

}
\keyword{cultural}
\keyword{ethnobotany,}
\keyword{report}
\keyword{use}
\keyword{value,}
