% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnoChordUser.R
\name{ethnoChordUser}
\alias{ethnoChordUser}
\title{Create a chord diagram of ethnobotany uses and species}
\source{
Whitney, C. W., Bahati, J., and Gebauer, J. (2018), Ethnobotany and agrobiodiversity; valuation of plants in the homegardens of southwestern Uganda. Ethnobiology Letters, 9(2), 90-100. <DOI:10.14237/ebl.9.2.2018.503.>
}
\usage{
ethnoChordUser(data)
}
\arguments{
\item{data}{is an ethnobotany data set with column 1 'informant' and 2 'sp_name' as row identifiers of informants and of species names respectively.
The rest of the columns are the identified ethnobotany use categories. The data should be populated with counts of uses per person (should be 0 or 1 values).}
}
\description{
This function allows you to create a chord diagram of informants and species uses for ethnobotany studies.
}
\examples{

ethnoChordUser(ethnobotanydata)

}
\keyword{cultural}
\keyword{ethnobotany,}
\keyword{report}
\keyword{use}
\keyword{value,}
