% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FCs.R
\name{FCs}
\alias{FCs}
\title{Frequency of citation (FC)}
\usage{
FCs(data)
}
\arguments{
\item{data}{is an ethnobotany data set with column 1 'informant' and 2 'sp_name' as row identifiers of informants and of species names respectively.
The rest of the columns are the identified ethnobotany use categories. The data should be populated with counts of uses per person (should be 0 or 1 values).}
}
\description{
This function allows you to calculate the frequency of citation (FC) per species.
}
\examples{

FCs(ethnobotanydata)

}
\keyword{ethnobotany,}
\keyword{number}
\keyword{of}
\keyword{quantitative}
\keyword{uses}
