% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIs.R
\name{RIs}
\alias{RIs}
\title{#Relative Importance Index (RI)}
\source{
Pardo-de-Santayana, M. 2003. Las plantas en la cultura tradicional de la antigua Merindad de Campoo. Ph.D. dissertation, Departamento de Biología, Facul- tad de Ciencias, Universidad Autónoma de Madrid, Spain.
}
\usage{
RIs(data)
}
\arguments{
\item{data}{is an ethnobotany data set with column 1 'informant' and 2 'sp_name' as row identifiers of informants and of species names respectively.
The rest of the columns are the identified ethnobotany use categories. The data should be populated with counts of uses per person (should be 0 or 1 values).}
}
\description{
This function allows you to calculate the relative importance index (RI) per species, published by Pardo-de-Santayana (2003).
}
\examples{

RIs(ethnobotanydata)

}
\keyword{ethnobotany,}
\keyword{importance}
\keyword{quantitative}
\keyword{relative}
