% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URsum.R
\name{URsum}
\alias{URsum}
\title{Sum of all Use Reports (UR) for all species}
\usage{
URsum(data)
}
\arguments{
\item{data}{is an ethnobotany data set with column 1 'informant' and 2 'sp_name' as row identifiers of informants and of species names respectively.
The rest of the columns are the identified ethnobotany use categories. The data should be populated with counts of uses per person (should be 0 or 1 values).}
}
\description{
Allows users to calculate the sum of all ethnobotany use reports (UR) for all species, a common metric for ethnobotany studies.
}
\examples{

#Use built-in ethnobotany data example
URsum(ethnobotanydata)

#Generate random dataset of three informants uses for four species
eb_data <- data.frame(replicate(10,sample(0:1,20,rep=TRUE)))
names(eb_data) <- gsub(x = names(eb_data), pattern = "X", replacement = "Use_")  
eb_data$informant<-sample(c('User_1', 'User_2', 'User_3'), 20, replace=TRUE)
eb_data$sp_name<-sample(c('sp_1', 'sp_2', 'sp_3', 'sp_4'), 20, replace=TRUE)
URsum(eb_data)

}
\keyword{ethnobotany}
\keyword{ethnobotany,}
\keyword{quantitative}
\keyword{report,}
\keyword{use}
