% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnobotanyChord.R
\name{ethnobotanyChord}
\alias{ethnobotanyChord}
\title{Chord diagram of informants and species uses}
\source{
Whitney, Cory W., Joseph Bahati, and J. Gebauer. 2018. “Ethnobotany and Agrobiodiversity; Valuation of Plants in the Homegardens of Southwestern Uganda.” Ethnobiology Letters 9 (2): 90–100.
}
\usage{
ethnobotanyChord(data)
}
\arguments{
\item{data}{is an ethnobotany data set with column 1 'informant' and 2 'sp_name' as row identifiers of informants and of species names respectively.
The rest of the columns are the identified ethnobotany use categories. The data should be populated with counts of uses per person (should be 0 or 1 values).}
}
\description{
Creates a chord diagram of informants and species uses for ethnobotany studies. For more on the circlize package see Zuguang Gu's 'Circular Visualization in R' <https://jokergoo.github.io/circlize_book/book/>
}
\section{Warning}{


Identification for informants and species must be listed by the names 'informant' and 'sp_name' respectively in the data set.
The rest of the columns represent the identified ethnobotany use categories. These data should be populated with counts of uses per person (should be 0 or 1 values).
}

\examples{

#Use built-in ethnobotany data example
ethnobotanyChord(ethnobotanydata)

#Generate random dataset of three informants uses for four species

eb_data <- data.frame(replicate(10,sample(0:1,20,rep=TRUE)))
names(eb_data) <- gsub(x = names(eb_data), pattern = "X", replacement = "Use_")  
eb_data$informant <- sample(c('User_1', 'User_2', 'User_3'), 20, replace=TRUE)
eb_data$sp_name <- sample(c('sp_1', 'sp_2', 'sp_3', 'sp_4'), 20, replace=TRUE)

ethnobotanyChord(eb_data)

}
\keyword{#'}
\keyword{'informant'}
\keyword{'sp_name'}
\keyword{(bottom}
\keyword{(top}
\keyword{@return}
\keyword{Chord}
\keyword{To}
\keyword{arith}
\keyword{change}
\keyword{data}
\keyword{diagram}
\keyword{dplyr}
\keyword{each}
\keyword{figure}
\keyword{for}
\keyword{function.}
\keyword{graphs}
\keyword{half)}
\keyword{in}
\keyword{logic}
\keyword{math}
\keyword{methods}
\keyword{misc}
\keyword{names}
\keyword{per}
\keyword{related}
\keyword{rename}
\keyword{set.}
\keyword{survey}
\keyword{the}
\keyword{to}
\keyword{try}
\keyword{use}
\keyword{using}
\keyword{variable}
