% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFCs.R
\name{RFCs}
\alias{RFCs}
\title{Relative Frequency of Citation (RFC)}
\source{
Tardio, J., and M. Pardo-de-Santayana, 2008. Cultural Importance Indices: A Comparative Analysis Based on the Useful Wild Plants of Southern Cantabria (Northern Spain) 1. Economic Botany, 62(1), 24-39. <https://doi.org/10.1007/s12231-007-9004-5>
}
\usage{
RFCs(data)
}
\arguments{
\item{data}{is an ethnobotany data set with column 1 'informant' and 2 'sp_name' as row identifiers of informants and of species names respectively.
The rest of the columns are the identified ethnobotany use categories. The data should be populated with counts of uses per person (should be 0 or 1 values).}
}
\description{
This function allows you to calculate the relative frequency of citation (RFC) per species published by Pardo-de-Santayana (2003).
}
\examples{
RFCs(ethnobotanydata)

}
\keyword{ethnobotany,}
\keyword{number}
\keyword{of}
\keyword{quantitative}
\keyword{uses}
