% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URs.R
\name{URs}
\alias{URs}
\title{Calculate the use report (UR) per species}
\usage{
URs(data)
}
\arguments{
\item{data}{is an ethnobotany data set with column 1 'informant' and 2 'sp_name' as row identifiers of informants and of species names respectively.
The rest of the columns are the identified ethnobotany use categories. The data should be populated with counts of uses per person (should be 0 or 1 values).}
}
\description{
This function allows you to calculate the use report (UR) per secies, a common metric for ethnobotany studies.
}
\examples{

URs(ethnobotanydata)

}
\keyword{cultural}
\keyword{ethnobotany,}
\keyword{report}
\keyword{use}
\keyword{value,}
